requirements.txt
Django==1.11.7
Pillow>=4.2.1
django-model-utils==3.0.0
djangorestframework==3.7.3
reportlab>=3.4.0
pip==9.0.1
html2pdf==0.1
